requirements.txt
python_version=3.9

scheduler==0.8.5
newspaper4k==0.9.2
mysql-connector-python==8.3.0
feedparser==6.0.11
python-dotenv==1.0.1
